% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_overlap.R
\name{.summarizeJASPARMotifs}
\alias{.summarizeJASPARMotifs}
\title{Summarize JASPAR motifs as taken from ArchR}
\usage{
.summarizeJASPARMotifs(motifs)
}
\arguments{
\item{motifs}{JASPAR motifs}
}
\value{
list of motifs and a summary dataframe
}
\description{
Summarize JASPAR motifs as taken from ArchR
}
