% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron_motifs.R
\name{intron_motifs}
\alias{intron_motifs}
\title{Match motifs against an annotation}
\usage{
intron_motifs(
  JASPAR,
  gff,
  genome = NULL,
  counts = TRUE,
  species = NULL,
  collection = "CORE",
  width = 7L,
  cutoff = 5e-05,
  bg = "even",
  kmer = 2L,
  percent_var = 99
)
}
\arguments{
\item{JASPAR}{A JASPAR SQLite file such as JASPAR2024::JASPAR2024()@db}

\item{gff}{GFF or GTF file to be read in by rtracklayer::readGFF}

\item{genome}{Genome string e.g. BSgenome.Hsapiens.UCSC.hg38}

\item{counts}{Whether to use counts or binary matches}

\item{species}{JASPAR species parameter. By default will be set to the BSgenome species}

\item{collection}{JASPAR collection}

\item{width}{Width for motifmatchr}

\item{cutoff}{P-value cutoff for motifmatchr}

\item{bg}{Background frequency for motifmatchr}

\item{kmer}{K-mer length to use for covariate adjustment}

\item{percent_var}{Cumulative percent variance captured of PCA of K-mers to use as a covariate}

\item{upstream}{Integer upstream}

\item{downstream}{Downstream base pairs of gene}
}
\value{
SummarizedExperiment object of genes by TF matrix of counts and associated metadata
}
\description{
Match motifs against an annotation
}
