% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifGLMFit.R
\name{motif_fit}
\alias{motif_fit}
\title{Fit Motifs}
\usage{
motif_fit(
  motif_se,
  feat_mat = NULL,
  covariates = NULL,
  use_expected = TRUE,
  size_factors = "poscounts",
  verbose = TRUE,
  model = "glmGamPoi",
  BPPARAM = bpparam(),
  ...
)
}
\description{
Fit Motifs
}
