% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_overlap.R
\name{motif_overlap}
\alias{motif_overlap}
\title{Motif overlap of genomic ranges}
\usage{
motif_overlap(
  peakSet,
  jaspar,
  species = "Homo sapiens",
  collection = "CORE",
  BSgenome = NULL,
  width = 7,
  cutoff = 5e-05,
  bg = "even",
  counts = FALSE
)
}
\arguments{
\item{peakSet}{GenomicRanges}

\item{jaspar}{JASPAR2024::JASPAR2024 or similar object}

\item{species}{Species to select from JASPAR}

\item{collection}{Collection from JASPAR}

\item{BSgenome}{BSgenome object to match peaks against. NULL means hg38}

\item{width}{Width for motifmatchr. Default=7}

\item{cutoff}{P-value cutoff for motifmatchr. Default=5e-05}
}
\value{
RangedSummarizedExperiment of each motif's enrichment
}
\description{
Motif overlap of genomic ranges
}
