% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test.R
\name{score_test.default}
\alias{score_test.default}
\title{Score test for default (numeric)}
\usage{
\method{score_test}{default}(X, Z, Res, Mu, BPPARAM = bpparam())
}
\description{
Score test for default (numeric)
}
