% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test.R
\name{score_test.DESeqDataSet}
\alias{score_test.DESeqDataSet}
\title{Score test for DESeqDataSet}
\usage{
\method{score_test}{DESeqDataSet}(dds, Z, BPPARAM = bpparam())
}
\description{
Score test for DESeqDataSet
}
