% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test.R
\name{score_test}
\alias{score_test}
\title{Score test for a fitted object.
Passing a BPPARAM and Z matrix (named) are required arguments.
The first argument can be a glmGamPoi object, DGEGLM object, or DESeqDataSet object.}
\usage{
score_test(X, ...)
}
\description{
Score test for a fitted object.
Passing a BPPARAM and Z matrix (named) are required arguments.
The first argument can be a glmGamPoi object, DGEGLM object, or DESeqDataSet object.
}
